Created CRUD for MbikqyresITemave, aligned it with the structure of my previous CRUDs
Created asynchronous methods to improve the performance and scalability, and also used AutoMapper to reduce the lines of code by automating the mapping between domain models and DTOs. Fixed the methods so that the userId is extracted from the token generated upon user login, and fixed the issue with Departamenti foreign key, for it to be saved in the MbikqyresITemave table correctly based on the name the user provides.