Final changes to User,Departamenti,Publikimi & Licensat CRUDs
Fixed the methods so that the userId is extracted from the token generated upon user login, and fixed the issues with Institucioni & Departamenti foreign keys, for them to be saved in the tables correctly based on the info the user provides.
Removed the AddUser method, since the user is created upon registration, and fixed the Update method, which will be used to edit the profile of a user.