Created and tested the form for updating Edukimi
After touching the edit button for a specific education shown in PersonDetails, it redirects you to the page where you can update the information about that education. The methods used here are get edukimi by id (to get the info in the fields) and update edukimi by id, to update the info as needed.