Created and tested the form for updating PunaVullnetare
After touching the edit button for a specific pune vullnetare shown in PersonDetails, it redirects you to the page where you can update the information about that pune vullnetare. The methods used here are get puna vullnetare by id (to get the info in the fields) and update puna vullnetare by id, to update the info as needed.