Created and tested the form for updating MbikqyresITemave
After touching the edit button for a specific mbikqyres shown in PersonDetails, it redirects you to the page where you can update the information about that mbikqyres. The methods used here are get mbikqyres by id (to get the info in the fields) and update mbikqyres by id, to update the info as needed.