Added an initial template for CV and fetched some data
I added a basic template for the CV, which will later be customized according to our needs. I implemented the fetching in a way that we can navigate from the PersonDetails to the CV, with the education data as part of the route state. The data is accessed from the route state and displayed in the CV. The way the data is being displayed will be changed later on. Lastly, through the buttons the user will be able to download the CV in PDF or JSON format.