Implemented the hide functionality for info of the CV
Work in progress for the CV data & design, fixed a redirection problem in Edit files, and implemented the hide functionality through an eye button. When that button is touched, the specific info will be hidden in the CV. I fixed that only for Edukimi for the moment, but later all entities will have that possibility.